requirements.txt
azure-ai-projects>=1.0.0
azure-ai-agents>=1.1.0
azure-identity>=1.19.0
python-dotenv>=1.0.1
